+++ .vscode/settings.json
@@ -2,5 +2,8 @@
 {
     "deno.enable": true,
     "deno.autoFmtOnSave": true,
-    "deno.alwaysShowStatus": true
+    "deno.alwaysShowStatus": true,
+    "editor.codeActionsOnSave": {
+        "source.organizeImports": true
+      }
 }
