+++ main-core/components-registry/componentRegistry.util.ts
@@ -1,12 +1,12 @@
-import { ReflectUtils } from "../utils/reflectUtils.ts";
-import { Reflect } from "../reflectMetadata.ts";
+import { MandarineORMException } from "../../orm-core/core/exceptions/mandarineORMException.ts";
+import { MandarineRepository } from "../../orm-core/repository/mandarineRepository.ts";
+import { ApplicationContext } from "../application-context/mandarineApplicationContext.ts";
 import { ComponentExceptions } from "../exceptions/componentExceptions.ts";
 import { InvalidAnnotationError } from "../exceptions/invalidAnnotationError.ts";
-import { MandarineConstants } from "../mandarineConstants.ts";
-import { ApplicationContext } from "../application-context/mandarineApplicationContext.ts";
 import { Mandarine } from "../Mandarine.ns.ts";
-import { MandarineRepository } from "../../orm-core/repository/mandarineRepository.ts";
-import { MandarineORMException } from "../../orm-core/core/exceptions/mandarineORMException.ts";
+import { MandarineConstants } from "../mandarineConstants.ts";
+import { Reflect } from "../reflectMetadata.ts";
+import { ReflectUtils } from "../utils/reflectUtils.ts";
 
 /**
 * This class contains utils that are used in references related to the DI container
@@ -14,7 +14,7 @@ import { MandarineORMException } from "../../orm-core/core/exceptions/mandarineO
 export class ComponentsRegistryUtil {
 
     public static registerComponent(componentTarget: any, componentType: Mandarine.MandarineCore.ComponentTypes, configuration: any, index: number): any {
-        if((index != null)) throw new InvalidAnnotationError(InvalidAnnotationError.CLASS_ONLY_ANNOTATION, Mandarine.MandarineCore.ComponentTypes[componentType]);
+        if((index != null)) throw new InvalidAnnotationError(InvalidAnnotationError.CLASS_ONLY_ANNOTATION);
         
         let parentClassName: string = ReflectUtils.getClassName(componentTarget);
         let componentName = parentClassName;
@@ -22,7 +22,7 @@ export class ComponentsRegistryUtil {
         let componentsRegistry = ApplicationContext.getInstance().getComponentsRegistry();
 
         if(componentsRegistry.exist(componentName)) {
-            throw new ComponentExceptions(ComponentExceptions.EXISTENT_COMPONENT.replace("%component%", componentName), componentName);
+            throw new ComponentExceptions(ComponentExceptions.EXISTENT_COMPONENT.replace("%component%", componentName));
         } else {
             Reflect.defineMetadata(`${MandarineConstants.REFLECTION_MANDARINE_COMPONENT}:${Mandarine.MandarineCore.ComponentTypes[componentType].toLowerCase()}:${componentName}`, {
                 componentName: componentName,
