+++ main-core/components/repository-component/repositoryComponent.ts
@@ -1,4 +1,3 @@
-import { MandarineException } from "../../exceptions/mandarineException.ts";
 import { Mandarine } from "../../Mandarine.ns.ts";
 
 /**
