+++ main-core/dependency-injection/di.util.ts
@@ -1,5 +1,5 @@
-import { Reflect } from "../reflectMetadata.ts";
 import { MandarineConstants } from "../mandarineConstants.ts";
+import { Reflect } from "../reflectMetadata.ts";
 import { ReflectUtils } from "../utils/reflectUtils.ts";
 import { DI } from "./di.ns.ts";
