+++ main-core/dependency-injection/diFactory.ts
@@ -1,12 +1,12 @@
-import { Mandarine, ComponentsRegistry, ViewModel } from "../../mod.ts";
-import { DI } from "./di.ns.ts";
-import { ReflectUtils } from "../utils/reflectUtils.ts";
-import { Reflect } from "../reflectMetadata.ts";
-import { MandarineConstants } from "../mandarineConstants.ts";
-import { RoutingUtils } from "../../mvc-framework/core/utils/mandarine/routingUtils.ts";
-import { HttpUtils } from "../utils/httpUtils.ts";
-import { ApplicationContext } from "../application-context/mandarineApplicationContext.ts";
 import { Cookies } from "../../deps.ts";
+import { ComponentsRegistry, Mandarine, ViewModel } from "../../mod.ts";
+import { RoutingUtils } from "../../mvc-framework/core/utils/mandarine/routingUtils.ts";
+import { ApplicationContext } from "../application-context/mandarineApplicationContext.ts";
+import { MandarineConstants } from "../mandarineConstants.ts";
+import { Reflect } from "../reflectMetadata.ts";
+import { HttpUtils } from "../utils/httpUtils.ts";
+import { ReflectUtils } from "../utils/reflectUtils.ts";
+import { DI } from "./di.ns.ts";
 
 export class DependencyInjectionFactory {
