+++ main-core/exceptions/routingException.ts
@@ -3,8 +3,8 @@ export class RoutingException extends Error {
     public static EXISTENT_ACTION: string = "The action cannot be created because one already exists. Routing Methods must not have the same name.";
     public static NONVALID_COMPONENT: string = "The action cannot be created because the component is not a valid controller.";
   
-    constructor(public message: string, public actionName: string) {
-      super(message + " ~ Object name: " + actionName);
+    constructor(public message: string) {
+      super(message);
       this.name = "RoutingException";
       this.stack = (this).stack;
     }
