+++ mvc-framework/engine/mandarineMvcFrameworkStarter.ts
@@ -1,13 +1,13 @@
-import { ApplicationContext } from "../../main-core/application-context/mandarineApplicationContext.ts";
-import { ControllerComponent } from "../core/internal/components/routing/controllerContext.ts";
-import { requestResolver, middlewareResolver } from "../core/internal/components/routing/routingResolver.ts";
-import { Log } from "../../logger/log.ts";
-import { SessionMiddleware } from "../core/middlewares/sessionMiddleware.ts";
-import { MiddlewareComponent } from "../../main-core/components/middleware-component/middlewareComponent.ts";
-import { Mandarine } from "../../main-core/Mandarine.ns.ts";
-import { getMandarineConfiguration } from "../../main-core/configuration/getMandarineConfiguration.ts";
 import { Router } from "../../deps.ts";
+import { Log } from "../../logger/log.ts";
+import { ApplicationContext } from "../../main-core/application-context/mandarineApplicationContext.ts";
+import { MiddlewareComponent } from "../../main-core/components/middleware-component/middlewareComponent.ts";
+import { getMandarineConfiguration } from "../../main-core/configuration/getMandarineConfiguration.ts";
+import { Mandarine } from "../../main-core/Mandarine.ns.ts";
+import { ControllerComponent } from "../core/internal/components/routing/controllerContext.ts";
+import { middlewareResolver, requestResolver } from "../core/internal/components/routing/routingResolver.ts";
 import { WebMVCConfigurer } from "../core/internal/configurers/webMvcConfigurer.ts";
+import { SessionMiddleware } from "../core/middlewares/sessionMiddleware.ts";
 
 /**
  * This class works as the MVC engine and it is responsible for the initialization & behavior of HTTP requests.
