+++ orm-core/connectors/postgreSQLConnector.ts
@@ -1,8 +1,8 @@
-import { Mandarine } from "../../main-core/Mandarine.ns.ts";
-import { Client as PostgresClient, Pool } from "https://deno.land/x/postgres/mod.ts";
 import { PoolClient } from "https://deno.land/x/postgres/client.ts";
-import { QueryResult, QueryConfig } from "https://deno.land/x/postgres/query.ts";
+import { Pool } from "https://deno.land/x/postgres/mod.ts";
+import { QueryConfig, QueryResult } from "https://deno.land/x/postgres/query.ts";
 import { Log } from "../../logger/log.ts";
+import { Mandarine } from "../../main-core/Mandarine.ns.ts";
 import { MandarineORMException } from "../core/exceptions/mandarineORMException.ts";
 
 export interface PostgresConnectorInterface extends Mandarine.ORM.Connection.Connector {
