+++ orm-core/entity-registry/entityRegistry.ts
@@ -1,7 +1,7 @@
+import { MandarineConstants } from "../../main-core/mandarineConstants.ts";
+import { Reflect } from "../../main-core/reflectMetadata.ts";
 import { ReflectUtils } from "../../main-core/utils/reflectUtils.ts";
 import { Mandarine } from "../../mod.ts";
-import { Reflect } from "../../main-core/reflectMetadata.ts";
-import { MandarineConstants } from "../../main-core/mandarineConstants.ts";
 
 /**
  * This class represents the registry where all the entities are added in order for mandarine to work with them at and after mandarine compile time.
