+++ orm-core/repository/mandarineRepository.ts
@@ -1,5 +1,5 @@
-import { Mandarine } from "../../main-core/Mandarine.ns.ts";
 import { ApplicationContext } from "../../main-core/application-context/mandarineApplicationContext.ts";
+import { Mandarine } from "../../main-core/Mandarine.ns.ts";
 
 export type RepositoryModeler = {
     instance: any,
